+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 
 - **fix**: mkdir in root
 - **feat**: try to implement auto focus
+- **feat**: suppress backend data refresh
 
 ## 0.5.5